+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
-## 0.1.0 (25-Sep-2020)
+## 0.4.0 (29-Sep-2020)
+- Basic version of Actor5e.js to create an appropriate structure
 
+## 0.3.0 (27-Sep-2020)
+- Fix traversal and create flat (fieldName, value) dictionary
+
+## 0.2.0 (27-Sep-2020)
+- Read XML and traverse imperfectly
+
+## 0.1.0 (25-Sep-2020)
 - **Created**
 -   Open an XFDF file from your local computer